pyproject.toml
[tool.poetry]
name = "rd-studio-backend"
version = "0.1.0"
description = "Django backend for RD Studio - handles image operations, bookings, and studio management"
authors = ["Praveen Maurya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.5"
djangorestframework = "^3.16.1"
pillow = "^11.3.0"
python-decouple = "^3.8"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
PyJWT = "^2.8.0"
djangorestframework-simplejwt = "^5.3.0"
boto3 = "^1.35.0"
requests = "^2.32.5"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
